server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lark==1.2.2
lxml==5.3.2
MarkupSafe==3.0.2
pdfkit==1.0.0
pillow==11.1.0
playwright==1.51.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydyf==0.11.0
pyee==12.1.1
pyparsing==3.2.3
pyphen==0.17.2
python-dotenv==1.1.0
reportlab==4.3.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
weasyprint==65.0
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
zipp==3.21.0
zopfli==0.2.3.post1
